requirements.txt
altgraph==0.17.3
pefile==2023.2.7
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.7
python-vlc==3.0.18122
pywin32-ctypes==0.2.2
tk==0.1.0
